Added More Validations and Logic to get minimum Weight
Added validations for input string and implemented the logic to get min weights

+++ WLBApplication/Data/IPlatesRepository.cs
@@ -6,6 +6,6 @@ namespace WLPBlatesManager.Model
     public interface IPlatesRepository
     {
         public Task<IEnumerable<Plate>> GetAllPlates();
-        public Task<Plate> GetPlate(double weight);
+        public Task<Plate> GetPlate(decimal weight);
     }
 }
