+++ WLBApplication/Model/Plate.cs
@@ -9,9 +9,9 @@ namespace WLPBlatesManager.Model
     public class Plate
     {
         public string Name { get; set; }
-        public double weight { get; set; }
+        public decimal weight { get; set; }
 
         [JsonIgnore]
-        public double pairedWeight { get { return weight * 2; } }
+        public decimal pairedWeight { get { return weight * 2; } }
     }
 }
